chore: separate entry points from util scripts
